requirements.txt
fastapi[standard]
python-jose
sqlalchemy
psycopg2-binary
python-dotenv
passlib
huaweicloudsdkcore
huaweicloudsdkfrs